Time: 0 ms (100%), Space: 40.8 MB (32.56%) - LeetHub

+++ 0191-number-of-1-bits/0191-number-of-1-bits.java
@@ -1,11 +1,11 @@
 public class Solution {
     
     public int hammingWeight(int n) {
-        int a=0;
+        short a=0;
         while(n!=0){           
             a+=(n&1);
             n>>=1;
             }
-        return a;
+        return (int)a;
     }
 }
